requirements.txt
streamlit==1.34.0
openai-whisper
torch
torchaudio
soundfile